moved more commands into bot/commands
