+++ bot/commands/avatar.py
@@ -1,0 +1,7 @@
+from ..betterbot import Member
+
+name = 'avatar'
+
+
+async def run(message, member: Member):
+	await message.channel.send(member.avatar_url)
