+++ bot/commands/b.py
@@ -1,0 +1,5 @@
+name = 'b'
+
+
+async def run(message):
+	await message.channel.send('I like french bread')
