+++ bot/commands/forum.py
@@ -1,0 +1,7 @@
+name = 'forum'
+aliases = ['forums', 'f']
+pad_none = False
+
+
+async def run(message):
+	await message.send('Forum commands: **!forums user (username)**')
