+++ bot/commands/wow.py
@@ -1,0 +1,6 @@
+name = 'wow'
+
+
+async def run(message):
+	await message.author.send('https://im-a-puzzle.com/look_i_m_a_puzzle_4XZqEhin.puzzle')
+	await message.delete()
